RS_Checkpoint_2/transaction-service/requirements.txt
fastapi
uvicorn
requests
python-jose
python-dotenv
